Add files via upload
Running the .bat file will cause it to access your PowerShell automatically and start running the program in the background, so be careful about that. (IT IS NOT A VIRUS) If you just want to view the code and file structure, don't touch the .bat file. You can view the map at http://localhost:8080 if you do run it.